# Reduce work done by RuntimeHelpers.TryGetHashCode on Mono