Following up from #81380 which implemented some fast paths for Mono's sync blocks in managed code.

* First correct `ObjectHeader.TryGetHashCode` to return true when it finds the hash code in the inflated lock word. This matches [the unmanaged implementation](https://github.com/dotnet/runtime/blob/907d0946162e302800495cce312c54f0e659db8c/src/mono/mono/metadata/monitor.c#L615-L621).
* Change `RuntimeHelpers.TryGetHashCode` to no longer make an icall if `ObjectHeader.TryGetHashCode` returns false. The icall does the same checks as `ObjectHeader.TryGetHashCode`. (see above unmanaged link).
* Remove the now unused icall `InternalTryGetHashCode`.