@AustinWise 2023-09-05
@lambdageek Could you take a look and see if this changes make sense? Also do you have a benchmark you would recommend running, perhaps this one?: