@lambdageek 2023-09-05
> @lambdageek Could you take a look and see if this changes make sense? Also do you have a benchmark you would recommend running, perhaps this one?: https://gist.github.com/lambdageek/c53949741489cac862d7918f2132b5a9

Yea that's the benchmark that I focused on.  As I recall, I primarily tried desktop JIT builds of Mono, and also WASM builds.

I wouldn't expect your change to introduce a perf regression. AFAICT it's removing code in more cases - either you return `true` more often, or you return false but avoid an icall.