@AustinWise 2023-09-09
I added the following two benchmarks to your set of benchmark to confirm that these changes have a positive effect:

```c#
    [Benchmark]
    public void GetHashCodeInflatedObject()
    {
        RuntimeHelpers.GetHashCode(inflatedLockObject);
    }

    private readonly ConditionalWeakTable<object, object> mWeakTable = new ()
    {
        {new object(), new object()},
    };

    [Benchmark]
    public bool TryGetNonExistentValue()
    {
        return mWeakTable.TryGetValue(inflatedLockObject, out object? _);
    }
```

Here are the results:

```
BenchmarkDotNet v0.13.8, macOS Ventura 13.5.2 (22G91) [Darwin 22.6.0]
Apple M2, 1 CPU, 8 logical and 8 physical cores
.NET SDK 8.0.100-preview.7.23376.3
  [Host]     : .NET 8.0.0 (8.0.23.37506), Arm64 RyuJIT AdvSIMD
  Job-RDANPG : .NET 9.0.0 (42.42.42.42424) using MonoVM, Arm64 ArmBase
  Job-OEETBD : .NET 9.0.0 (42.42.42.42424) using MonoVM, Arm64 ArmBase
```

| Method                       | Job        | Toolchain | Mean      | Error     | StdDev    | Ratio |
|----------------------------- |----------- |---------- |----------:|----------:|----------:|------:|
| LockPulseUnlockFlatObjNoHash | Job-RDANPG | PR        | 28.919 ns | 0.1478 ns | 0.1234 ns |  1.00 |
| LockPulseUnlockFlatObjNoHash | Job-OEETBD | base      | 28.840 ns | 0.2225 ns | 0.2081 ns |  1.00 |
|                              |            |           |           |           |           |       |
| LockUnlockFlatObjNoHash      | Job-RDANPG | PR        | 17.684 ns | 0.1008 ns | 0.0943 ns |  0.98 |
| LockUnlockFlatObjNoHash      | Job-OEETBD | base      | 17.992 ns | 0.0630 ns | 0.0590 ns |  1.00 |
|                              |            |           |           |           |           |       |
| GetHashCodeNoLock            | Job-RDANPG | PR        |  3.505 ns | 0.0057 ns | 0.0050 ns |  0.98 |
| GetHashCodeNoLock            | Job-OEETBD | base      |  3.566 ns | 0.0013 ns | 0.0011 ns |  1.00 |
|                              |            |           |           |           |           |       |
| LockUnlockInflatedObject     | Job-RDANPG | PR        | 19.526 ns | 0.0720 ns | 0.0674 ns |  1.00 |
| LockUnlockInflatedObject     | Job-OEETBD | base      | 19.490 ns | 0.1665 ns | 0.1558 ns |  1.00 |
|                              |            |           |           |           |           |       |
| GetHashCodeInflatedObject    | Job-RDANPG | PR        |  2.907 ns | 0.0218 ns | 0.0204 ns |  0.33 |
| GetHashCodeInflatedObject    | Job-OEETBD | base      |  8.834 ns | 0.0199 ns | 0.0177 ns |  1.00 |
|                              |            |           |           |           |           |       |
| TryGetNonExistentValue       | Job-RDANPG | PR        |  8.876 ns | 0.0152 ns | 0.0142 ns |  0.63 |
| TryGetNonExistentValue       | Job-OEETBD | base      | 14.091 ns | 0.0046 ns | 0.0041 ns |  1.00 |


I also ran with the `MONO_ENV_OPTIONS:--interp=-tiering` option and confirmed there was no effect on these benchmarks, as expected.